# openapi_schema.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "CarbonGrid-AI",
                 "description":  "Carbon-aware middleware for local LLM inference",
                 "version":  "0.1.0"
             },
    "paths":  {
...
                  "/generate":  {
                                    "post":  {
                                                 "summary":  "Generate",
                                                 "description":  "Generate text with carbon-aware precision selection.\n\nPipeline:\n1. Extract classifier features from prompt\n2. Get safety probability from C.1.1 classifier\n3. Get current carbon intensity\n4. Run Decision Engine\n5. Execute inference at selected precision\n6. Return response with full telemetry",
                                                 "operationId":  "generate_generate_post",
                                                 "requestBody":  {
                                                                     "content":  {
                                                                                     "application/json":  {
                                                                                                              "schema":  {
                                                                                                                             "$ref":  "#/components/schemas/GenerationRequest"
                                                                                                                         }
                                                                                                          }
                                                                                 },
                                                                     "required":  true
                                                                 },
                                                 "responses":  {
                                                                   "200":  {
                                                                               "description":  "Successful Response",
                                                                               "content":  {
                                                                                               "application/json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/GenerationResponse"
                                                                                                                                   }
                                                                                                                    }
                                                                                           }
                                                                           },
                                                                   "422":  {
                                                                               "description":  "Validation Error",
                                                                               "content":  {
                                                                                               "application/json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                   }
                                                                                                                    }
                                                                                           }
                                                                           }
                                                               }
                                             }
                                },
                  "/decision/test":  {
                                         "post":  {
                                                      "summary":  "Test Decision",
                                                      "description":  "Test the Decision Engine directly with custom input.",
                                                      "operationId":  "test_decision_decision_test_post",
                                                      "requestBody":  {
                                                                          "content":  {
                                                                                          "application/json":  {
                                                                                                                   "schema":  {
                                                                                                                                  "additionalProperties":  true,
                                                                                                                                  "type":  "object",
                                                                                                                                  "title":  "Input Data"
                                                                                                                              }
                                                                                                               }
                                                                                      },
                                                                          "required":  true
                                                                      },
                                                      "responses":  {
                                                                        "200":  {
                                                                                    "description":  "Successful Response",
                                                                                    "content":  {
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {

                                                                                                                                        }
                                                                                                                         }
                                                                                                }
                                                                                },
                                                                        "422":  {
                                                                                    "description":  "Validation Error",
                                                                                    "content":  {
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                        }
                                                                                                                         }
                                                                                                }
                                                                                }
                                                                    }
                                                  }
                                     },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "GenerationResponse":  {
                                                                  "properties":  {
                                                                                     "response":  {
                                                                                                      "type":  "string",
                                                                                                      "title":  "Response",
                                                                                                      "description":  "Generated text response"
                                                                                                  },
                                                                                     "carbon_grid":  {
                                                                                                         "additionalProperties":  true,
                                                                                                         "type":  "object",
                                                                                                         "title":  "Carbon Grid",
                                                                                                         "description":  "CarbonGrid decision and telemetry data"
                                                                                                     }
                                                                                 },
                                                                  "type":  "object",
                                                                  "required":  [
                                                                                   "response"
                                                                               ],
                                                                  "title":  "GenerationResponse",
                                                                  "description":  "Response from text generation with CarbonGrid telemetry."
                                                              },
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
...
                                   }
                   }
}